Add my examples
