+++ .env.sample
@@ -1,11 +1,8 @@
-# API_HOST can be either azure, ollama, openai, or github:
+# API_HOST can be either azure, openai, or github:
 API_HOST=azure
 # Configure for Azure:
 AZURE_OPENAI_ENDPOINT=https://YOUR-AZURE-OPENAI-SERVICE-NAME.openai.azure.com/openai/v1
 AZURE_OPENAI_CHAT_DEPLOYMENT=YOUR-AZURE-DEPLOYMENT-NAME
-# Configure for Ollama:
-OLLAMA_ENDPOINT=http://localhost:11434/v1
-OLLAMA_MODEL=llama3.1
 # Configure for OpenAI.com:
 OPENAI_API_KEY=YOUR-OPENAI-KEY
 OPENAI_MODEL=gpt-3.5-turbo
