+++ .gitattributes
@@ -1,0 +1,3 @@
+* text=auto
+*.sh text eol=lf
+*.ps1 text eol=crlf
