+++ examples/expenses.csv
@@ -1,0 +1,2 @@
+date,amount,category,description,payment_method
+2026-02-08,1200.0,gadget,Laptop purchase,visa
