+++ examples/openai_tool_calling.py
@@ -1,0 +1,89 @@
+import json
+import os
+
+import azure.identity
+import openai
+from dotenv import load_dotenv
+from rich import print
+
+# Configure OpenAI client based on environment
+load_dotenv(override=True)
+API_HOST = os.getenv("API_HOST", "github")
+
+if API_HOST == "azure":
+    token_provider = azure.identity.get_bearer_token_provider(
+        azure.identity.DefaultAzureCredential(), "https://cognitiveservices.azure.com/.default"
+    )
+    client = openai.OpenAI(
+        base_url=f"{os.environ['AZURE_OPENAI_ENDPOINT']}/openai/v1/",
+        api_key=token_provider,
+    )
+    MODEL_NAME = os.environ["AZURE_OPENAI_CHAT_DEPLOYMENT"]
+elif API_HOST == "github":
+    client = openai.OpenAI(base_url="https://models.github.ai/inference", api_key=os.environ["GITHUB_TOKEN"])
+    MODEL_NAME = os.getenv("GITHUB_MODEL", "openai/gpt-5-mini")
+else:
+    client = openai.OpenAI(api_key=os.environ["OPENAI_API_KEY"])
+    MODEL_NAME = os.environ.get("OPENAI_MODEL", "gpt-5-mini")
+
+
+def lookup_weather(city_name: str | None = None, zip_code: str | None = None) -> dict:
+    """Lookup the weather for a given city name or zip code."""
+    return {
+        "city_name": city_name,
+        "zip_code": zip_code,
+        "weather": "sunny",
+        "temperature": 75,
+    }
+
+
+tools = [
+    {
+        "type": "function",
+        "function": {
+            "name": "lookup_weather",
+            "description": "Lookup the weather for a given city name or zip code.",
+            "parameters": {
+                "type": "object",
+                "properties": {
+                    "city_name": {
+                        "type": "string",
+                        "description": "The city name",
+                    },
+                    "zip_code": {
+                        "type": "string",
+                        "description": "The zip code",
+                    },
+                },
+                "strict": True,
+                "additionalProperties": False,
+            },
+        },
+    }
+]
+
+messages = [
+    {"role": "system", "content": "You are a weather chatbot."},
+    {"role": "user", "content": "is it sunny in berkeley CA?"},
+]
+response = client.chat.completions.create(
+    model=MODEL_NAME,
+    messages=messages,
+    tools=tools,
+    tool_choice="auto",
+)
+
+print(f"Response from {MODEL_NAME} on {API_HOST}: \n")
+
+# Now actually call the function as indicated
+if response.choices[0].message.tool_calls:
+    tool_call = response.choices[0].message.tool_calls[0]
+    function_name = tool_call.function.name
+    arguments = json.loads(tool_call.function.arguments)
+
+    if function_name == "lookup_weather":
+        messages.append(response.choices[0].message)
+        result = lookup_weather(**arguments)
+        messages.append({"role": "tool", "tool_call_id": tool_call.id, "content": str(result)})
+        response = client.chat.completions.create(model=MODEL_NAME, messages=messages, tools=tools)
+        print(response.choices[0].message.content)
