+++ examples/spanish/agent_tool.py
@@ -1,0 +1,77 @@
+import asyncio
+import logging
+import os
+import random
+from typing import Annotated
+
+from agent_framework import ChatAgent
+from agent_framework.openai import OpenAIChatClient
+from azure.identity.aio import DefaultAzureCredential, get_bearer_token_provider
+from dotenv import load_dotenv
+from pydantic import Field
+from rich import print
+from rich.logging import RichHandler
+
+# Setup logging
+handler = RichHandler(show_path=False, rich_tracebacks=True, show_level=False)
+logging.basicConfig(level=logging.WARNING, handlers=[handler], force=True, format="%(message)s")
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+# Configurar el cliente para usar Azure OpenAI, GitHub Models o OpenAI
+load_dotenv(override=True)
+API_HOST = os.getenv("API_HOST", "github")
+
+async_credential = None
+if API_HOST == "azure":
+    async_credential = DefaultAzureCredential()
+    token_provider = get_bearer_token_provider(async_credential, "https://cognitiveservices.azure.com/.default")
+    client = OpenAIChatClient(
+        base_url=f"{os.environ['AZURE_OPENAI_ENDPOINT']}/openai/v1/",
+        api_key=token_provider,
+        model_id=os.environ["AZURE_OPENAI_CHAT_DEPLOYMENT"],
+    )
+elif API_HOST == "github":
+    client = OpenAIChatClient(
+        base_url="https://models.github.ai/inference",
+        api_key=os.environ["GITHUB_TOKEN"],
+        model_id=os.getenv("GITHUB_MODEL", "openai/gpt-5-mini"),
+    )
+else:
+    client = OpenAIChatClient(api_key=os.environ["OPENAI_API_KEY"], model_id=os.environ.get("OPENAI_MODEL", "gpt-5-mini"))
+
+
+def get_weather(
+    city: Annotated[str, Field(description="Nombre completo de la ciudad")],
+) -> dict:
+    """Devuelve datos meteorológicos para una ciudad: temperatura y descripción."""
+    logger.info(f"Obteniendo el clima para {city}")
+    if random.random() < 0.05:
+        return {
+            "temperature": 72,
+            "description": "Soleado",
+        }
+    else:
+        return {
+            "temperature": 60,
+            "description": "Lluvioso",
+        }
+
+
+agent = ChatAgent(
+    chat_client=client,
+    instructions="Eres un agente informativo. Responde a las preguntas con alegría.",
+    tools=[get_weather],
+)
+
+
+async def main():
+    response = await agent.run("¿Cómo está el clima hoy en San Francisco?")
+    print(response.text)
+
+    if async_credential:
+        await async_credential.close()
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
