+++ examples/spanish/openai_tool_calling.py
@@ -1,0 +1,89 @@
+import json
+import os
+
+import azure.identity
+import openai
+from dotenv import load_dotenv
+from rich import print
+
+# Configurar el cliente para usar Azure OpenAI, GitHub Models o OpenAI
+load_dotenv(override=True)
+API_HOST = os.getenv("API_HOST", "github")
+
+if API_HOST == "azure":
+    token_provider = azure.identity.get_bearer_token_provider(
+        azure.identity.DefaultAzureCredential(), "https://cognitiveservices.azure.com/.default"
+    )
+    client = openai.OpenAI(
+        base_url=f"{os.environ['AZURE_OPENAI_ENDPOINT']}/openai/v1/",
+        api_key=token_provider,
+    )
+    MODEL_NAME = os.environ["AZURE_OPENAI_CHAT_DEPLOYMENT"]
+elif API_HOST == "github":
+    client = openai.OpenAI(base_url="https://models.github.ai/inference", api_key=os.environ["GITHUB_TOKEN"])
+    MODEL_NAME = os.getenv("GITHUB_MODEL", "openai/gpt-5-mini")
+else:
+    client = openai.OpenAI(api_key=os.environ["OPENAI_API_KEY"])
+    MODEL_NAME = os.environ.get("OPENAI_MODEL", "gpt-5-mini")
+
+
+def lookup_weather(city_name: str | None = None, zip_code: str | None = None) -> dict:
+    """Consulta el clima para un nombre de ciudad o código postal dado."""
+    return {
+        "city_name": city_name,
+        "zip_code": zip_code,
+        "weather": "soleado",
+        "temperature": 24,
+    }
+
+
+tools = [
+    {
+        "type": "function",
+        "function": {
+            "name": "lookup_weather",
+            "description": "Consulta el clima para un nombre de ciudad o código postal dado.",
+            "parameters": {
+                "type": "object",
+                "properties": {
+                    "city_name": {
+                        "type": "string",
+                        "description": "El nombre de la ciudad",
+                    },
+                    "zip_code": {
+                        "type": "string",
+                        "description": "El código postal",
+                    },
+                },
+                "strict": True,
+                "additionalProperties": False,
+            },
+        },
+    }
+]
+
+messages = [
+    {"role": "system", "content": "Eres un chatbot del clima. Responde en español."},
+    {"role": "user", "content": "¿Está soleado en Madrid?"},
+]
+response = client.chat.completions.create(
+    model=MODEL_NAME,
+    messages=messages,
+    tools=tools,
+    tool_choice="auto",
+)
+
+print(f"Respuesta de {MODEL_NAME} en {API_HOST}: \n")
+
+# Ahora ejecutar la función según lo indicado
+if response.choices[0].message.tool_calls:
+    tool_call = response.choices[0].message.tool_calls[0]
+    function_name = tool_call.function.name
+    arguments = json.loads(tool_call.function.arguments)
+
+    if function_name == "lookup_weather":
+        messages.append(response.choices[0].message)
+        result = lookup_weather(**arguments)
+        messages.append({"role": "tool", "tool_call_id": tool_call.id, "content": str(result)})
+        response = client.chat.completions.create(model=MODEL_NAME, messages=messages, tools=tools)
+        print(response.choices[0].message.content)
