+++ infra/write_dot_env.sh
