+++ requirements.txt
@@ -6,5 +6,6 @@ rich
 dotenv-azd
 aiohttp
 faker
+fastmcp
 agent-framework-core @ git+https://github.com/microsoft/agent-framework.git@98cd72839e4057d661a58092a3b013993264d834#subdirectory=python/packages/core
 agent-framework-devui @ git+https://github.com/microsoft/agent-framework.git@98cd72839e4057d661a58092a3b013993264d834#subdirectory=python/packages/devui
